refactor: extract assistant api client helpers
