+++ docs/repository-maturity-plan.md
@@ -137,6 +137,8 @@ scripts/
 26. `ensureHostedHistorySeeded(...)` 与 `runHostedAgentStepForSession(...)` 开始复用 orchestrator helper，hosted session/history/step 串联调度继续从页面中抽离。
 27. 新增 `src/lib/assistant-api-meta.js`，抽离 assistant API 类型归一化、鉴权要求、提示文案与输入占位元数据。
 28. `assistant.js` 开始复用 assistant API meta helper，页面层不再内联维护 API 类型说明与占位规则。
+29. 新增 `src/lib/assistant-api-client.js`，抽离 assistant API base URL 规整、鉴权头构造与重试请求逻辑。
+30. `assistant.js` 开始复用 assistant API client helper，assistant 页与底层 API 客户端细节进一步解耦。
 
 ## 风险与回滚建议
 - 风险：`chat.js` 仍然较大，后续继续拆分时容易影响事件时序。
